Extract any macros
